refactor: moved most functions to do with database to database.py
